Add rollup
Building to dist/ three versions of the package, with names specified by `main`, `module`, and `browser` in package.json.

For this particular project, this is definitely overkill. This is a Node module, and so will likely only be in a CommonJS environment. But it was a good learning exercise.

When we publish to npm, we will only include the files in dist/, assuming the `files` entry in package.json works as intended. (I think I am unable to test this locally, as including local packages just creates a symlink in node_modules/ to the full package dir.)